Fix binary gcc toolchain, point rust symlink to correct location.
The toolchain by default ships with a rust symlink, which points to
/home/redox/rust, and we don't know if that is actually where the redox
repo resides for the user. As a fix, we package the rust/ folder in the
redox src into it's own derivation, and then just replace to symlink to
there instead. E.G. we go from $gcc_toolchain/lib/rustlib/src/rust ->
/home/bender/rust to $gcc_toolchain/lib/rustlib/src/rust ->
$redox_rust_folder_derivation